requirements.txt
distribute==0.6.28
envoy==0.0.2
requests==0.14.2
websocket-client==0.8.0
wsgiref==0.1.2
